refactor: 예외처리 패키지 분리 (#70)
